requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
dj-stripe==2.4.3
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
djongo==1.3.4
filelock==3.0.12
geomet==0.2.1.post1
idna==2.10
jsonfield==3.1.0
mongoengine==0.23.0
Pillow==8.1.2
psycopg2==2.8.6
pymongo==3.11.3
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.2.4
stripe==2.56.0
urllib3==1.26.4
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6

channels~=3.0.3